#11561 Import from CSV: target column type detect fix (integers, reals)

+++ plugins/org.jkiss.dbeaver.data.transfer/src/org/jkiss/dbeaver/tools/transfer/database/DatabaseTransferUtils.java
@@ -345,7 +345,7 @@ public class DatabaseTransferUtils {
             return DATA_TYPE_UNKNOWN;
         }
         try {
-            Integer.parseInt(value);
+            Long.parseLong(value);
             return DATA_TYPE_INTEGER;
         } catch (NumberFormatException ignored) {
         }
